Port the x2timer-ng mode1 code from the pic16f1778 implementation.
Compared with the 1778 implementation, we have:
New pin assignments for the smaller 20-pin chip.
An entry in Flash.h for the __16F1769 chip.
Also, we don't use an op-amp to buffer the DAC1 output.

At this commit, the firmware build succeeds but
we have not yet tried to program the MCU.