Introduce a simple firmware-controller trigger mode
so that we can return to the CLC trigger function later.

Also, introduce use of second DAC, clean up reading of analog channels,
clean up code around the use of the virtual registers,
add empty functions to hold the code for the delayed-trigger modes and
add a description of the commands and registers.